# Add ParseStrict and MustParseStrict methods.