These are methods like `Parse` and `MustParse`, but validate that their inputs contain only valid base32 characters. On invalid input, they return a new `ErrInvalidCharacters` error rather than producing an undefined ULID.

A comparison benchmark:
```
$ go test -v -run=^$ -benchtime 10s -bench BenchmarkParse ./...
goos: darwin
goarch: amd64
pkg: github.com/oklog/ulid
BenchmarkParse-8         	500000000	        26.1 ns/op	 994.49 MB/s
BenchmarkParseStrict-8   	500000000	        39.2 ns/op	 662.87 MB/s
PASS
ok  	github.com/oklog/ulid	39.221s
?   	github.com/oklog/ulid/cmd/ulid	[no test files]
```

This is an alternate fix for https://github.com/oklog/ulid/issues/21. I originally approached this issue by adding this validation to `Parse` and `UnmarshalText` (https://github.com/oklog/ulid/pull/22), but there was some concern about the performance impact.

Fixes https://github.com/oklog/ulid/issues/21.