@mattmoyer 2018-07-13
One question about the interface here as long as we're adding a new method: should this require that the input is _exactly_ the same format returned by `.String()`?

For example, it currently accepts lowercase base32. Should it reject those?