@tsenart 2018-07-14
> For example, it currently accepts lowercase base32. Should it reject those?

I don't think it should reject those.