@mattmoyer 2018-07-16
@tsenart: thanks for the review!

I updated to address review comments, please take a look at your leisure. The only other new bit I added was `MustParseStrict()` which is just like `MustParse()` but with `ParseStrict()` underneath.