@mattmoyer 2018-07-16
@tsenart I updated to split out some internal validation functions that are shared between the parsing functions:

- `validateSize()`
- `validateOverflow()`
- `validateCharacterSet()`
- `decodeInto()`

Unfortunately, some of these functions are too big for Go to want to inline them, which hurts performance:
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkParse-8           25.6          30.4          +18.75%
BenchmarkParseStrict-8     35.7          44.2          +23.81%

benchmark                  old MB/s     new MB/s     speedup
BenchmarkParse-8           1014.82      854.11       0.84x
BenchmarkParseStrict-8     728.90       588.59       0.81x
```

In addition to the function call overhead, I think there are some bounds checks that can't be elided because the code isn't inlined.

The only other approach I can think of would be to copy-paste the code and manually inline it at both locations, which feels messy to me.