@tsenart 2018-07-16
We're dealing here directly with the trade-off of performance for maintainability and usability. @peterbourgon: I know you said you didn't want to make a judgment call on this, but I'd like your thoughts on this again.